backend_app/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
blinker==1.9.0
boto3==1.36.0
celery==5.4.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
django-rest-framework-mongoengine==3.4.1
djangorestframework-simplejwt==5.4.0
django-environ==0.11.2
dnspython==2.7.0
# grpcio==1.69.0
# grpcio-tools==1.69.0
# protobuf==5.29.3
graphene==3.4.3
graphene-django==3.2.2
graphene-mongo==0.4.4
gunicorn==23.0.0
h11==0.14.0
kombu==5.4.2
mongoengine==0.29.1
packaging==24.2
prompt_toolkit==3.0.48
pymongo==4.10.1
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==0.19.1
redis==5.0.7
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
tenacity==8.0.1
tzdata==2024.2
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
websockets==12.0
